@@ -1,7 +1,7 @@
 package net.kwmt27.codesearch.entity;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.format.DateFormat;
 import android.text.style.ClickableSpan;
 import android.view.View;
