@@ -1,7 +1,7 @@
 package net.kwmt27.codesearch.presenter.search;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 
 import net.kwmt27.codesearch.ModelLocator;
